Currencies: move currency formatting basics into a utility.
