+++ client/payment-details/summary/index.js
@@ -3,12 +3,11 @@
 /**
  * External dependencies
  */
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { dateI18n } from '@wordpress/date';
 import { Card } from '@woocommerce/components';
-import Currency, { getCurrencyData } from '@woocommerce/currency';
 import moment from 'moment';
-import { find, get } from 'lodash';
+import { get } from 'lodash';
 
 /**
  * Internal dependencies.
@@ -20,30 +19,8 @@ import HorizontalList from 'components/horizontal-list';
 import Loadable, { LoadableBlock } from 'components/loadable';
 import riskMappings from 'components/risk-level/strings';
 import OrderLink from 'components/order-link';
+import { isZeroDecimalCurrency, getCurrency } from 'utils/currency';
 import './style.scss';
-
-const currencyData = getCurrencyData();
-
-/**
- * Gets wc-admin Currency for the given currency code
- *
- * @param {String} currencyCode Currency code
- *
- * @return {Currency} Currency object
- */
-const getCurrency = ( currencyCode ) => {
-	const currency = find( currencyData, { code: currencyCode.toUpperCase() } );
-	if ( currency ) {
-		return new Currency( currency );
-	}
-	window.console.warn(
-		sprintf(
-			'"%s" is not supported by @woocommerce/currency, falling back to "USD"',
-			currencyCode
-		)
-	);
-	return new Currency();
-};
 
 const placeholderValues = {
 	net: 0,
@@ -101,31 +78,9 @@ const composePaymentSummaryItems = ( { charge } ) =>
 const PaymentDetailsSummary = ( { charge = {}, isLoading } ) => {
 	const formatCurrency = ( amount, currency ) => {
 		const currencyCode = currency || charge.currency || 'USD';
-		const zeroDecimalCurrencies = [
-			'bif',
-			'clp',
-			'djf',
-			'gnf',
-			'jpy',
-			'kmf',
-			'krw',
-			'mga',
-			'pyg',
-			'rwf',
-			'ugx',
-			'vnd',
-			'vuv',
-			'xaf',
-			'xof',
-			'xpf',
-		];
-		const isZeroDecimalCurrency = zeroDecimalCurrencies.includes(
-			currencyCode.toLowerCase()
-		);
-		if ( isZeroDecimalCurrency ) {
+		if ( isZeroDecimalCurrency( currencyCode ) ) {
 			amount *= 100;
 		}
-
 		return getCurrency( currencyCode ).formatCurrency( amount / 100 );
 	};
 	const { net, fee, refunded } = charge.amount
